enforce exection order for core and addon etl procedures
